[feature] SpringRacingCarToJPA - Cars 객체 생성
